reorganize packages into versions
