--- src/main/java/org/flux/store/api/ReducerBlock.java
@@ -1,5 +1,0 @@
-package org.flux.store.api;
-
-public interface ReducerBlock<T extends State> extends Reducer<T> {
-    String getType();
-}
