+++ src/main/java/org/flux/store/api/exceptions/AsyncNotificationException.java
@@ -1,4 +1,4 @@
-package org.flux.store.api;
+package org.flux.store.api.exceptions;
 
 public class AsyncNotificationException extends Exception{
     public AsyncNotificationException(String message) {super(message);}
