+++ src/main/java/org/flux/store/api/v1/Reducer.java
@@ -1,4 +1,4 @@
-package org.flux.store.api;
+package org.flux.store.api.v1;
 
 @FunctionalInterface
 public interface Reducer<T extends State> {
