+++ src/main/java/org/flux/store/api/v1/Store.java
@@ -1,4 +1,4 @@
-package org.flux.store.api;
+package org.flux.store.api.v1;
 
 import java.util.function.Consumer;
