+++ src/main/java/org/flux/store/api/v2/Slice.java
@@ -1,6 +1,8 @@
-package org.flux.store.api;
+package org.flux.store.api.v2;
 
-import java.lang.reflect.Type;
+import org.flux.store.api.exceptions.InvalidActionException;
+import org.flux.store.api.v1.State;
+
 import java.util.function.Consumer;
 
 public interface Slice<T extends State> {
