+++ src/test/java/org/flux/store/tests/reflection/SetNameReducer.java
@@ -1,8 +1,8 @@
 package org.flux.store.tests.reflection;
 
-import org.flux.store.api.Action;
-import org.flux.store.api.ReducerBlock;
-import org.flux.store.api.AutoStore;
+import org.flux.store.api.v1.Action;
+import org.flux.store.api.v3.ReducerBlock;
+import org.flux.store.api.v3.AutoStore;
 import org.flux.store.tests.domain.UserProfile;
 
 @AutoStore("MyStore")
