+++ src/test/java/org/flux/store/tests/utils/CombinedReducerTest.java
@@ -1,8 +1,8 @@
-package org.flux.store.tests;
+package org.flux.store.tests.utils;
 
-import org.flux.store.api.Action;
-import org.flux.store.api.Reducer;
-import org.flux.store.main.DuxStore;
+import org.flux.store.api.v1.Action;
+import org.flux.store.api.v1.Reducer;
+import org.flux.store.main.v1.DuxStore;
 import org.flux.store.utils.Utilities;
 import org.flux.store.tests.domain.Author;
 import org.flux.store.tests.domain.Book;
