+++ src/test/java/org/flux/store/tests/v2/MiddlewareTest.java
@@ -1,9 +1,9 @@
-package org.flux.store.tests;
+package org.flux.store.tests.v2;
 
-import org.flux.store.api.Action;
-import org.flux.store.api.Middleware;
-import org.flux.store.api.Reducer;
-import org.flux.store.main.DuxStore;
+import org.flux.store.api.v1.Action;
+import org.flux.store.api.v2.Middleware;
+import org.flux.store.api.v1.Reducer;
+import org.flux.store.main.v1.DuxStore;
 import org.flux.store.utils.Utilities;
 import org.flux.store.tests.domain.UserProfile;
 import org.junit.jupiter.api.BeforeEach;
